requirements.txt
pyelftools>=0.29
openpyxl>=3.1.0
